fix: adjustments in remove comments
