+++ src/comments/ParticipantsSection.tsx
