+++ src/comments/PendingMessageBubble.tsx
